run fibonacci directly instead of invoking handler
